+++ app/locales/en/translations.ts
@@ -716,6 +716,19 @@ export default {
         contact_us: 'Contact us',
         title: 'Institutions',
         search_placeholder: 'Search institutions',
+        dashboard: {
+            select_default: 'All Departments',
+            users_list: {
+                name: 'Name',
+                department: 'Department',
+                projects: 'Projects',
+                public_projects: 'Public Projects',
+                private_projects: 'Private Projects',
+            },
+            users_connected_panel: 'SSO Users Connected',
+            projects_panel: 'Total Projects',
+            departments_panel: 'Departments',
+        },
     },
     registration: {
         project_alert: 'This registration is a frozen, non-editable version of <a class="link-solid" href="/{{projectId}}/">this project</a>',
@@ -1122,6 +1135,10 @@ export default {
                 },
                 form: {
                     title: 'Fill Out Registration Form',
+                    next: 'Next',
+                    back: 'Back',
+                    save_draft: 'Save draft',
+                    new_registration: 'New Registration',
                 },
                 review: {
                     title: 'Review Registration Before Submitting',
@@ -1387,6 +1404,29 @@ export default {
             cancelButton: 'Cancel',
             selectAll: 'Select All',
             clearAll: 'Clear All',
+        },
+        finalizeRegistrationModal: {
+            title: 'Almost done...',
+            notice: `Please keep in mind that:
+                <ul>
+                    <li>
+                        Registrations cannot be modified or deleted once completed.
+                    </li>
+                    <li>
+                        The content and version history of Wiki and OSF Storage will be copied to the registration.
+                    </li>
+                    <li>
+                        This project contains links to other projects.These links will be copied into your registration, but the
+                        projects that they link to will not be registered.If you wish to register the linked projects, they must be
+                        registered separately.Learn more about links.
+                    </li>
+                </ul>`,
+            immediateOption: 'Make registration public immediately',
+            embargoOption: 'Enter registration into embargo',
+            createDoi: 'Create DOI',
+            submit: 'Submit',
+            back: 'Back',
+            datePlaceholder: 'Choose embargo end date',
         },
     },
     meetings: {