+++ app/models/registration.ts
@@ -3,7 +3,9 @@ import { computed } from '@ember-decorators/object';
 import { buildValidations, validator } from 'ember-cp-validations';
 import DS from 'ember-data';
 
-import { ResponseValue } from 'ember-osf-web/packages/registration-schema';
+import DraftRegistrationModel from 'ember-osf-web/models/draft-registration';
+import { RegistrationResponse } from 'ember-osf-web/packages/registration-schema';
+
 import CommentModel from './comment';
 import ContributorModel from './contributor';
 import InstitutionModel from './institution';
@@ -11,8 +13,6 @@ import NodeModel from './node';
 import RegistrationProviderModel from './registration-provider';
 import RegistrationSchemaModel, { RegistrationMetadata } from './registration-schema';
 import UserModel from './user';
-
-export type RegistrationResponses = Record<string, ResponseValue>;
 
 export enum RegistrationState {
     Embargoed = 'Embargoed',
@@ -55,12 +55,11 @@ export default class RegistrationModel extends NodeModel.extend(Validations) {
     @attr('fixstring') registrationSupplement?: string;
     @attr('fixstring') articleDoi!: string | null;
     @attr('object') registeredMeta!: RegistrationMetadata;
-    @attr('object') registrationResponses!: RegistrationResponses;
+    @attr('object') registrationResponses!: RegistrationResponse;
 
     // Write-only attributes
-    @attr('fixstring') draftRegistration?: string;
-    @attr('fixstring') registrationChoice?: 'immediate' | 'embargo';
-    @attr('date') liftEmbargo?: Date;
+    @attr('array') includedNodeIds?: string[];
+    @attr('boolean') createDoi?: boolean;
 
     @computed(
         'withdrawn', 'embargoed', 'public', 'pendingRegistrationApproval',
@@ -105,6 +104,10 @@ export default class RegistrationModel extends NodeModel.extend(Validations) {
 
     @hasMany('institution', { inverse: 'registrations' })
     affiliatedInstitutions!: DS.PromiseManyArray<InstitutionModel> | InstitutionModel[];
+
+    // Write-only relationships
+    @belongsTo('draft-registration', { inverse: null })
+    draftRegistration!: DraftRegistrationModel;
 
     registrationStateMap(): Record<RegistrationState, boolean> {
         const {