+++ app/packages/registration-schema/validations.ts
@@ -2,14 +2,14 @@ import { assert } from '@ember/debug';
 import { ValidationObject, ValidatorFunction } from 'ember-changeset-validations';
 import { validateLength, validatePresence } from 'ember-changeset-validations/validators';
 import translations from 'ember-osf-web/locales/en/translations';
-import { PageResponse } from 'ember-osf-web/packages/registration-schema';
+import { RegistrationResponse } from 'ember-osf-web/packages/registration-schema';
 import { SchemaBlockGroup } from 'ember-osf-web/packages/registration-schema/schema-block-group';
 import { validateResponseFormat } from 'ember-osf-web/validators/validate-response-format';
 
 // TODO: find a way to use i18n to translate error messages
 
 export function buildValidation(groups: SchemaBlockGroup[]) {
-    const ret: ValidationObject<PageResponse> = {};
+    const ret: ValidationObject<RegistrationResponse> = {};
     groups.forEach((group: SchemaBlockGroup) => {
         // only validating groups with actual inputs and not groups that are headings/labels only
         if (group.inputBlock) {