+++ mirage/scenarios/default.ts
@@ -40,6 +40,11 @@ function registrationScenario(
         index: 0,
     });
 
+    server.create('draft-registration', {
+        id: 'decaf',
+        registrationSchema: server.schema.registrationSchemas.find('testSchema'),
+    });
+
     registerNodeMultiple(
         server,
         registrationNode,
@@ -161,7 +166,7 @@ function dashboardScenario(server: Server, currentUser: ModelInstance<User>) {
     // NOTE: Some institutions are already created by this point
     server.createList('institution', 20);
     // Create a specific institution to test institutional dashboard with; should be ID 29 at this point
-    server.create('institution', { id: 'has-users' }, 'withInstitutionalUsers');
+    server.create('institution', { id: 'has-users' }, 'withInstitutionalUsers', 'withStatSummary');
 }
 
 function forksScenario(server: Server, currentUser: ModelInstance<User>) {